corrections

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -10,11 +10,16 @@ from sqlalchemy.orm import sessionmaker
 from model_state import State
 
 if __name__ == "__main__":
-    engine = create_engine("mysql+mysqldb://{}:{}@localhost/{}".format(
-        sys.argv[1], sys.argv[2], sys.argv[3]), pool_pre_ping=True)
-    session_maker = sessionmaker(bind=engine)
-    session = session_maker()
-    states = session.query(State).filter(State.name.like('%a%')).all()
-    for state in states:
+    if len(sys.argv) != 4:
+        print("Usage: {} <username> <password> <database>".format(sys.argv[0]))
+        sys.exit(1)
+    engine = create_engine('mysql+mysqldb://{}:{}@localhost/{}'
+                           .format(sys.argv[1], sys.argv[2], sys.argv[3]),
+                           pool_pre_ping=True)
+    Session = sessionmaker(bind=engine)
+    session = Session()
+    states_with_a = session.query(State).filter(State.name.like('%a%')).all()
+    for state in states_with_a:
         session.delete(state)
     session.commit()
+    session.close()
